requirements.txt
# Core dependencies
pathway>=0.13.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Optional (if using other features)
# torch>=2.0.0  # Required by sentence-transformers (usually auto-installed)
